Change create_username_dist.sh to loop through each directory in given directory
Added a for loop to go through each directory in the given directory and
put all of the usernames into a "listofusers.txt" and then sorted it,
counted it, and then used awk to format the main content for
username.html (which is the list of data.addRow[]s). Then combined the
header, footer, and middle content into a file called
username_dist.html.